join ffmpeg commands into one
Before we handled reading and outputing the audio stream separaretly. For larger (not that big tho) files with for example 5:02 minutes length, this had buggy behavior as the output was truncated to some arbritrary amount (e.g. 2:27 or 4:03). Joining the command not only makes the process faster, but also fixes this bug and makes the code cleaner.